+++ hypothesis-python/src/hypothesis/core.py
@@ -334,12 +334,6 @@ def execute_explicit_examples(state, wrapped_test, arguments, kwargs):
 
         with local_settings(state.settings):
             fragments_reported = []
-
-            def report_buffered():
-                for f in fragments_reported:
-                    report(f)
-                del fragments_reported[:]
-
             try:
                 with with_reporter(fragments_reported.append):
                     state.execute_once(
@@ -354,17 +348,20 @@ def execute_explicit_examples(state, wrapped_test, arguments, kwargs):
                 # a saved failure when other arguments are supplied by e.g. pytest.
                 # See https://github.com/HypothesisWorks/hypothesis/issues/2125
                 pass
-            except BaseException:
-                report_buffered()
-                raise
+            except BaseException as err:
+                # In order to support reporting of multiple failing examples, we yield
+                # each of the (report text, error) pairs we find back to the top-level
+                # runner.  This also ensures that user-facing stack traces have as few
+                # frames of Hypothesis internals as possible.
+                yield (fragments_reported, err.with_traceback(get_trimmed_traceback()))
+                if state.settings.report_multiple_bugs:
+                    continue
+                break
 
-            if current_verbosity() >= Verbosity.verbose:
-                prefix = "Falsifying example"
-                assert fragments_reported[0].startswith(prefix)
-                fragments_reported[0] = (
-                    "Trying example" + fragments_reported[0][len(prefix) :]
-                )
-                report_buffered()
+            assert fragments_reported[0].startswith("Falsifying example")
+            verbose_report(fragments_reported[0].replace("Falsifying", "Trying", 1))
+            for f in fragments_reported[1:]:
+                verbose_report(f)
 
 
 def get_random_for_wrapped_test(test, wrapped_test):
@@ -576,9 +573,10 @@ class StateForActualGivenExecution:
                                         printer.text(",")
                                         printer.breakable()
 
-                                    # We need to make sure to print these in the argument order for
-                                    # Python 2 and older versionf of Python 3.5. In modern versions
-                                    # this isn't an issue because kwargs is ordered.
+                                    # We need to make sure to print these in the
+                                    # argument order for Python 2 and older versions
+                                    # of Python 3.5. In modern versions this isn't
+                                    # an issue because kwargs is ordered.
                                     arg_order = {
                                         v: i
                                         for i, v in enumerate(
@@ -887,7 +885,28 @@ class HypothesisHandle:
     """
 
     inner_test = attr.ib()
-    fuzz_one_input = attr.ib()
+    _get_fuzz_target = attr.ib()
+
+    @property
+    def fuzz_one_input(
+        self,
+    ) -> Callable[[Union[bytes, bytearray, memoryview, BinaryIO]], Optional[bytes]]:
+        """Run the test as a fuzz target, driven with the `buffer` of bytes.
+
+        Returns None if buffer invalid for the strategy, canonical pruned
+        bytes if the buffer was valid, and leaves raised exceptions alone.
+
+        Note: this feature is experimental and may change or be removed.
+        """
+        # Note: most users, if they care about fuzzer performance, will access the
+        # property and assign it to a local variable to move the attribute lookup
+        # outside their fuzzing loop / before the fork point.  We cache it anyway,
+        # so that naive or unusual use-cases get the best possible performance too.
+        try:
+            return self.__cached_target  # type: ignore
+        except AttributeError:
+            self.__cached_target = self._get_fuzz_target()
+            return self.__cached_target
 
 
 def given(
@@ -1047,8 +1066,28 @@ def given(
 
             # There was no @reproduce_failure, so start by running any explicit
             # examples from @example decorators.
-
-            execute_explicit_examples(state, wrapped_test, arguments, kwargs)
+            errors = list(
+                execute_explicit_examples(state, wrapped_test, arguments, kwargs)
+            )
+            with local_settings(state.settings):
+                if len(errors) > 1:
+                    # If we're not going to report multiple bugs, we would have
+                    # stopped running explicit examples at the first failure.
+                    assert state.settings.report_multiple_bugs
+                    for fragments, err in errors:
+                        for f in fragments:
+                            report(f)
+                        tb_lines = traceback.format_exception(
+                            type(err), err, err.__traceback__
+                        )
+                        report("".join(tb_lines))
+                    msg = "Hypothesis found %d failures in explicit examples."
+                    raise MultipleFailures(msg % (len(errors)))
+                elif errors:
+                    fragments, the_error_hypothesis_found = errors[0]
+                    for f in fragments:
+                        report(f)
+                    raise the_error_hypothesis_found
 
             # If there were any explicit examples, they all ran successfully.
             # The next step is to use the Conjecture engine to run the test on
@@ -1101,24 +1140,22 @@ def given(
                     )
                     raise the_error_hypothesis_found
 
-        def fuzz_one_input(
-            buffer: Union[bytes, bytearray, memoryview, BinaryIO]
-        ) -> Optional[bytes]:
-            """Run the test as a fuzz target, driven with the `buffer` of bytes.
-
-            Returns None if buffer invalid for the strategy, canonical pruned
-            bytes if the buffer was valid, and leaves raised exceptions alone.
-
-            Note: this feature is experimental and may change or be removed.
-            """
-            if isinstance(buffer, io.IOBase):
-                buffer = buffer.read()
-            if isinstance(buffer, (bytearray, memoryview)):
-                buffer = bytes(buffer)
-            assert isinstance(buffer, bytes)
-
+        def _get_fuzz_target() -> Callable[
+            [Union[bytes, bytearray, memoryview, BinaryIO]], Optional[bytes]
+        ]:
+            # Because fuzzing interfaces are very performance-sensitive, we use a
+            # somewhat more complicated structure here.  `_get_fuzz_target()` is
+            # called by the `HypothesisHandle.fuzz_one_input` property, allowing
+            # us to defer our collection of the settings, random instance, and
+            # reassignable `inner_test` (etc) until `fuzz_one_input` is accessed.
+            #
+            # We then share the performance cost of setting up `state` between
+            # many invocations of the target.  We explicitly force `deadline=None`
+            # for performance reasons, saving ~40% the runtime of an empty test.
             test = wrapped_test.hypothesis.inner_test
-            settings = wrapped_test._hypothesis_internal_use_settings
+            settings = Settings(
+                parent=wrapped_test._hypothesis_internal_use_settings, deadline=None
+            )
             random = get_random_for_wrapped_test(test, wrapped_test)
             _args, _kwargs, test_runner, search_strategy = process_arguments_to_given(
                 wrapped_test, (), {}, given_kwargs, argspec, settings,
@@ -1128,17 +1165,29 @@ def given(
             state = StateForActualGivenExecution(
                 test_runner, search_strategy, test, settings, random, wrapped_test,
             )
+            digest = function_digest(test)
 
-            data = ConjectureData.for_buffer(buffer)
-            try:
-                state.execute_once(data)
-            except (StopTest, UnsatisfiedAssumption):
-                return None
-            except BaseException:
-                if settings.database is not None:
-                    settings.database.save(function_digest(test), bytes(data.buffer))
-                raise
-            return bytes(data.buffer)
+            def fuzz_one_input(
+                buffer: Union[bytes, bytearray, memoryview, BinaryIO]
+            ) -> Optional[bytes]:
+                # This inner part is all that the fuzzer will actually run,
+                # so we keep it as small and as fast as possible.
+                if isinstance(buffer, io.IOBase):
+                    buffer = buffer.read()
+                assert isinstance(buffer, (bytes, bytearray, memoryview))
+                data = ConjectureData.for_buffer(buffer)
+                try:
+                    state.execute_once(data)
+                except (StopTest, UnsatisfiedAssumption):
+                    return None
+                except BaseException:
+                    if settings.database is not None:
+                        settings.database.save(digest, bytes(data.buffer))
+                    raise
+                return bytes(data.buffer)
+
+            fuzz_one_input.__doc__ = HypothesisHandle.fuzz_one_input.__doc__
+            return fuzz_one_input
 
         # After having created the decorated test function, we need to copy
         # over some attributes to make the switch as seamless as possible.
@@ -1159,7 +1208,7 @@ def given(
         wrapped_test._hypothesis_internal_use_reproduce_failure = getattr(
             test, "_hypothesis_internal_use_reproduce_failure", None
         )
-        wrapped_test.hypothesis = HypothesisHandle(test, fuzz_one_input)
+        wrapped_test.hypothesis = HypothesisHandle(test, _get_fuzz_target)
         return wrapped_test
 
     return run_test_as_given